Added Frontend Functionality
Added the ability for the user to choose a size for their input matricies. Also made it so the input matricies resize to fit the user's choice. Partially integrated with Backend code. Just need to hook up Matrix B to get inputs as well as display answers to the screen.