Masters Theorem, MagnitudeOrder updates
Master Theorem now only deals with one actual state, the recurrence relation received from the user. I also fixed an issue I found with the solver not detecting parentheses on the dValue. Added the functionality to send to Database.

Magnitude Order is now more selective to reduce waiting time of the user. Need to push to database